% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsl_common_vars_extractor.R
\name{print.common_adsl_vars}
\alias{print.common_adsl_vars}
\title{Print Method for Common ADSL Variables}
\usage{
\method{print}{common_adsl_vars}(x, ...)
}
\arguments{
\item{x}{Data frame returned by extract_common_adsl_variables}

\item{...}{Additional arguments passed to print}
}
\description{
Custom print method for the output of extract_common_adsl_variables
}
