% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsl_common_vars_extractor.R
\name{extract_variable_info_from_define}
\alias{extract_variable_info_from_define}
\title{Extract Variable Information from Define.xml}
\usage{
extract_variable_info_from_define(define_path)
}
\arguments{
\item{define_path}{Character string specifying the path to the define.xml file}
}
\value{
A data.frame with columns: Name, Dataset, DataType, Length, Description, Origin, etc.
}
\description{
This helper function parses a define.xml file to extract variable metadata
that can be used with extract_common_adsl_variables.
}
\examples{
\dontrun{
define_path <- system.file("define.xml", package = "adrgOS")
if (file.exists(define_path)) {
  var_info <- extract_variable_info_from_define(define_path)
  print(head(var_info))
}
}
}
